Fix VmCloud button handling when navigating from other objects
* Move controller code for Attach/Detach/LiveMigrate/Evacuate/Resize
  into `application_controller/ci_processing.rb` so that it
  is accessible to other controllers.

* Rewrite controller code to support being called from both
  explorer views and normal views.

* Add conditional code to views so that they'll render
  form submission buttons when not inside of an explorer view

* Add cases to `button` method in EmsCommonController and
  CloudVolumeController so that they can handle VmCloud buttons.
