+++ app/controllers/vm_cloud_controller.rb
@@ -11,409 +11,6 @@ class VmCloudController < ApplicationController
     @table_name ||= "vm_cloud"
   end
 
-  def resize
-    assert_privileges("instance_resize")
-    @record = find_by_id_filtered(VmOrTemplate, params[:id]) # Set the VM object
-    if @record.is_available?(:resize)
-      drop_breadcrumb(
-        :name => _("Reconfigure Instance '%{name}'") % {:name => @record.name},
-        :url  => "/vm_cloud/resize"
-      ) unless @explorer
-      @flavors = {}
-      unless @record.ext_management_system.nil?
-        @record.ext_management_system.flavors.each do |ems_flavor|
-          # include only flavors with root disks at least as big as the instance's current root disk.
-          if (ems_flavor != @record.flavor) && (ems_flavor.root_disk_size >= @record.flavor.root_disk_size)
-            @flavors[ems_flavor.name_with_details] = ems_flavor.id
-          end
-        end
-      end
-      @edit = {}
-      @edit[:new] ||= {}
-      unless @record.flavor.nil?
-        @edit[:new][:flavor] = @record.flavor.id
-      end
-      @edit[:key] = "vm_resize__#{@record.id}"
-      @edit[:vm_id] = @record.id
-      @edit[:explorer] = true if params[:action] == "x_button" || session.fetch_path(:edit, :explorer)
-      session[:edit] = @edit
-      @in_a_form = true
-      @refresh_partial = "vm_common/resize"
-    else
-      add_flash(_("Unable to reconfigure %{instance} \"%{name}\": %{details}") % {
-        :instance => ui_lookup(:table => 'vm_cloud'),
-        :name     => @record.name,
-        :details  => @record.is_available_now_error_message(:resize)}, :error)
-    end
-  end
-  alias instance_resize resize
-
-  def resize_vm
-    assert_privileges("instance_resize")
-    load_edit("vm_resize__#{params[:id]}")
-    flavor_id = @edit[:new][:flavor]
-    flavor = find_by_id_filtered(Flavor, flavor_id)
-    @record = VmOrTemplate.find_by_id(params[:id])
-
-    case params[:button]
-    when "cancel"
-      add_flash(_("Reconfigure of %{model} \"%{name}\" was cancelled by the user") % {
-        :model => ui_lookup(:table => "vm_cloud"), :name => @record.name})
-      @record = @sb[:action] = nil
-      replace_right_cell
-    when "submit"
-      if @record.is_available?(:resize)
-        begin
-          old_flavor = @record.flavor
-          @record.resize(flavor)
-          add_flash(_("Reconfiguring %{instance} \"%{name}\" from %{old_flavor} to %{new_flavor}") % {
-            :instance   => ui_lookup(:table => 'vm_cloud'),
-            :name       => @record.name,
-            :old_flavor => old_flavor.name,
-            :new_flavor => flavor.name})
-        rescue => ex
-          add_flash(_("Unable to reconfigure %{instance} \"%{name}\": %{details}") % {
-            :instance => ui_lookup(:table => 'vm_cloud'),
-            :name     => @record.name,
-            :details  => ex}, :error)
-        end
-      else
-        add_flash(_("Unable to reconfigure %{instance} \"%{name}\": %{details}") % {
-          :instance => ui_lookup(:table => 'vm_cloud'),
-          :name     => @record.name,
-          :details  => @record.is_avaiable_now_error_message(:resize)}, :error)
-      end
-      params[:id] = @record.id.to_s # reset id in params for show
-      @record = nil
-      @sb[:action] = nil
-      replace_right_cell
-    end
-  end
-
-  def resize_field_changed
-    return unless load_edit("vm_resize__#{params[:id]}")
-    @edit ||= {}
-    @edit[:new] ||= {}
-    @edit[:new][:flavor] = params[:flavor]
-    render :update do |page|
-      page << javascript_prologue
-      page.replace_html("main_div",
-                        :partial => "vm_common/resize") if %w(allright left right).include?(params[:button])
-      page << javascript_for_miq_button_visibility(true)
-      page << "miqSparkle(false);"
-    end
-  end
-
-  def live_migrate
-    assert_privileges("instance_live_migrate")
-    @record = find_by_id_filtered(VmOrTemplate, params[:id]) # Set the VM object
-    if @record.is_available?(:live_migrate) && !@record.ext_management_system.nil?
-      drop_breadcrumb(
-        :name => _("Live Migrate Instance '%{name}'") % {:name => @record.name},
-        :url  => "/vm_cloud/live_migrate"
-      ) unless @explorer
-      @in_a_form = true
-      @refresh_partial = "vm_common/live_migrate"
-    else
-      add_flash(_("Unable to live migrate %{instance} \"%{name}\": %{details}") % {
-        :instance => ui_lookup(:table => 'vm_cloud'),
-        :name     => @record.name,
-        :details  => @record.is_available_now_error_message(:live_migrate)}, :error)
-    end
-  end
-  alias_method :instance_live_migrate, :live_migrate
-
-  def live_migrate_form_fields
-    assert_privileges("instance_live_migrate")
-    @record = find_by_id_filtered(VmOrTemplate, params[:id])
-    hosts = []
-    unless @record.ext_management_system.nil?
-      # wrap in a rescue block in the event the connection to the provider fails
-      begin
-        connection = @record.ext_management_system.connect
-        current_hostname = connection.handled_list(:servers).find do |s|
-          s.name == @record.name
-        end.os_ext_srv_attr_hypervisor_hostname
-        # OS requires its own name for the host be used in the migrate API, so get the
-        # provider hostname from fog.
-        hosts = connection.hosts.select { |h| h.service_name == "compute" && h.host_name != current_hostname }.map do |h|
-          {:name => h.host_name, :id => h.host_name}
-        end
-      rescue
-        hosts = []
-      end
-    end
-    render :json => {
-      :hosts => hosts
-    }
-  end
-
-  def live_migrate_vm
-    assert_privileges("instance_live_migrate")
-    @record = find_by_id_filtered(VmOrTemplate, params[:id])
-
-    case params[:button]
-    when "cancel"
-      cancel_action(_("Live Migration of %{model} \"%{name}\" was cancelled by the user") % {
-        :model => ui_lookup(:table => 'vm_cloud'),
-        :name  => @record.name
-      })
-    when "submit"
-      if @record.is_available?(:live_migrate)
-        if params['auto_select_host'] == 'on'
-          hostname = nil
-        else
-          hostname = params[:destination_host]
-        end
-        block_migration = params[:block_migration]
-        disk_over_commit = params[:disk_over_commit]
-        begin
-          @record.live_migrate(
-            :hostname         => hostname,
-            :block_migration  => block_migration == 'on',
-            :disk_over_commit => disk_over_commit == 'on'
-          )
-          add_flash(_("Live Migrating %{instance} \"%{name}\"") % {
-            :instance => ui_lookup(:table => 'vm_cloud'),
-            :name     => @record.name})
-        rescue => ex
-          add_flash(_("Unable to live migrate %{instance} \"%{name}\": %{details}") % {
-            :instance => ui_lookup(:table => 'vm_cloud'),
-            :name     => @record.name,
-            :details  => get_error_message_from_fog(ex.to_s)}, :error)
-        end
-      else
-        add_flash(_("Unable to live migrate %{instance} \"%{name}\": %{details}") % {
-          :instance => ui_lookup(:table => 'vm_cloud'),
-          :name     => @record.name,
-          :details  => @record.is_available_now_error_message(:live_migrate)}, :error)
-      end
-      params[:id] = @record.id.to_s # reset id in params for show
-      @record = nil
-      @sb[:action] = nil
-      replace_right_cell
-    end
-  end
-
-  def attach
-    assert_privileges("instance_attach")
-    @volume_choices = {}
-    @record = @vm = find_by_id_filtered(VmCloud, params[:id])
-    @vm.cloud_tenant.cloud_volumes.each { |volume| @volume_choices[volume.name] = volume.id }
-
-    @in_a_form = true
-    drop_breadcrumb(
-      :name => _("Attach %{volume} to %{instance_model} \"%{instance_name}\"") % {
-        :volume         => ui_lookup(:table => 'cloud_volume'),
-        :instance_model => ui_lookup(:table => 'vm_cloud'),
-        :instance_name  => @vm.name
-      },
-      :url  => "/vm_cloud/attach")
-    @in_a_form = true
-    @refresh_partial = "vm_common/attach"
-  end
-  alias instance_attach attach
-
-  def detach
-    assert_privileges("instance_detach")
-    @volume_choices = {}
-    @record = @vm = find_by_id_filtered(VmCloud, params[:id])
-    attached_volumes = @vm.hardware.disks.select(&:backing).map(&:backing)
-    attached_volumes.each { |volume| @volume_choices[volume.name] = volume.id }
-    if attached_volumes.empty?
-      add_flash(_("%{instance_model} \"%{instance_name}\" has no attached %{volumes}") % {
-        :volumes        => ui_lookup(:tables => 'cloud_volumes'),
-        :instance_model => ui_lookup(:table => 'vm_cloud'),
-        :instance_name  => @vm.name})
-      render :update do |page|
-        page.replace("flash_msg_div", :partial => "layouts/flash_msg")
-      end
-    end
-
-    @in_a_form = true
-    drop_breadcrumb(
-      :name => _("Detach %{volume} from %{instance_model} \"%{instance_name}\"") % {
-        :volume         => ui_lookup(:table => 'cloud_volume'),
-        :instance_model => ui_lookup(:table => 'vm_cloud'),
-        :instance_name  => @vm.name
-      },
-      :url  => "/vm_cloud/detach")
-    @in_a_form = true
-    @refresh_partial = "vm_common/detach"
-  end
-  alias instance_detach detach
-
-  def attach_volume
-    assert_privileges("instance_attach")
-
-    @vm = find_by_id_filtered(VmCloud, params[:id])
-    case params[:button]
-    when "cancel"
-      cancel_action(_("Attaching %{volume_model} to %{instance_model} \"%{instance_name}\" was cancelled by the user") % {
-        :volume_model   => ui_lookup(:table => 'cloud_volume'),
-        :instance_model => ui_lookup(:table => 'vm_cloud'),
-        :instance_name  => @vm.name
-      })
-    when "attach"
-      volume = find_by_id_filtered(CloudVolume, params[:volume_id])
-      if volume.is_available?(:attach_volume)
-        begin
-          volume.raw_attach_volume(@vm.ems_ref, params[:device_path])
-          add_flash(_("Attaching %{volume} \"%{volume_name}\" to %{vm_name}") % {
-            :volume      => ui_lookup(:table => 'cloud_volume'),
-            :volume_name => volume.name,
-            :vm_name     => @vm.name})
-        rescue => ex
-          add_flash(_("Unable to attach %{volume} \"%{volume_name}\" to %{vm_name}: %{details}") % {
-            :volume      => ui_lookup(:table => 'cloud_volume'),
-            :volume_name => volume.name,
-            :vm_name     => @vm.name,
-            :details     => get_error_message_from_fog(ex)}, :error)
-        end
-      else
-        add_flash(_(volume.is_available_now_error_message(:attach_volume)), :error)
-      end
-      @breadcrumbs.pop if @breadcrumbs
-      session[:edit] = nil
-      @record = @sb[:action] = nil
-      replace_right_cell
-
-    end
-  end
-
-  def detach_volume
-    assert_privileges("instance_detach")
-
-    @vm = find_by_id_filtered(VmCloud, params[:id])
-    case params[:button]
-    when "cancel"
-      cancel_action(_("Detaching a %{volume} from %{instance_model} \"%{instance_name}\" was cancelled by the user") % {
-        :volume         => ui_lookup(:table => 'cloud_volume'),
-        :instance_model => ui_lookup(:table => 'vm_cloud'),
-        :instance_name  => @vm.name
-      })
-
-    when "detach"
-      volume = find_by_id_filtered(CloudVolume, params[:volume_id])
-      if volume.is_available?(:detach_volume)
-        begin
-          volume.raw_detach_volume(@vm.ems_ref)
-          add_flash(_("Detaching %{volume} \"%{volume_name}\" from %{vm_name}") % {
-            :volume      => ui_lookup(:table => 'cloud_volume'),
-            :volume_name => volume.name,
-            :vm_name     => @vm.name})
-        rescue => ex
-          add_flash(_("Unable to detach %{volume} \"%{volume_name}\" from %{vm_name}: %{details}") % {
-            :volume      => ui_lookup(:table => 'cloud_volume'),
-            :volume_name => volume.name,
-            :vm_name     => @vm.name,
-            :details     => get_error_message_from_fog(ex)}, :error)
-        end
-      else
-        add_flash(_(volume.is_available_now_error_message(:detach_volume)), :error)
-      end
-      @breadcrumbs.pop if @breadcrumbs
-      session[:edit] = nil
-      @record = @sb[:action] = nil
-      replace_right_cell
-    end
-  end
-
-  def cancel_action(message)
-    session[:edit] = nil
-    add_flash(message)
-    @record = @sb[:action] = nil
-    replace_right_cell
-  end
-
-  def evacuate
-    assert_privileges("instance_evacuate")
-    @record = find_by_id_filtered(VmOrTemplate, params[:id]) # Set the VM object
-    if @record.is_available?(:evacuate) && !@record.ext_management_system.nil?
-      drop_breadcrumb(
-        :name => _("Evacuate Instance '%{name}'") % {:name => @record.name},
-        :url  => "/vm_cloud/evacuate"
-      ) unless @explorer
-      @in_a_form = true
-      @refresh_partial = "vm_common/evacuate"
-    else
-      add_flash(_("Unable to evacuate %{instance} \"%{name}\": %{details}") % {
-        :instance => ui_lookup(:table => 'vm_cloud'),
-        :name     => @record.name,
-        :details  => @record.is_available_now_error_message(:evacuate)}, :error)
-    end
-  end
-  alias_method :instance_evacuate, :evacuate
-
-  def evacuate_form_fields
-    assert_privileges("instance_evacuate")
-    @record = find_by_id_filtered(VmOrTemplate, params[:id])
-    hosts = []
-    unless @record.ext_management_system.nil?
-      begin
-        connection = @record.ext_management_system.connect
-        current_hostname = connection.handled_list(:servers).find do |s|
-          s.name == @record.name
-        end.os_ext_srv_attr_hypervisor_hostname
-        hosts = connection.hosts.select { |h| h.service_name == "compute" && h.host_name != current_hostname }.map do |h|
-          {:name => h.host_name, :id => h.host_name}
-        end
-      rescue
-        hosts = []
-      end
-    end
-    render :json => {
-      :hosts => hosts
-    }
-  end
-
-  def evacuate_vm
-    assert_privileges("instance_evacuate")
-    @record = find_by_id_filtered(VmOrTemplate, params[:id])
-
-    case params[:button]
-    when "cancel"
-      cancel_action(_("Evacuation of %{model} \"%{name}\" was cancelled by the user") % {
-        :model => ui_lookup(:table => 'vm_cloud'),
-        :name  => @record.name
-      })
-    when "submit"
-      if @record.is_available?(:evacuate)
-        if params['auto_select_host'] == 'on'
-          hostname = nil
-        else
-          hostname = params[:destination_host]
-        end
-        on_shared_storage = params[:on_shared_storage] == 'on'
-        admin_password = on_shared_storage ? nil : params[:admin_password]
-        begin
-          @record.evacuate(
-            :hostname          => hostname,
-            :on_shared_storage => on_shared_storage,
-            :admin_password    => admin_password
-          )
-          add_flash(_("Evacuating %{instance} \"%{name}\"") % {
-            :instance => ui_lookup(:table => 'vm_cloud'),
-            :name     => @record.name})
-        rescue => ex
-          add_flash(_("Unable to evacuate %{instance} \"%{name}\": %{details}") % {
-            :instance => ui_lookup(:table => 'vm_cloud'),
-            :name     => @record.name,
-            :details  => get_error_message_from_fog(ex.to_s)}, :error)
-        end
-      else
-        add_flash(_("Unable to evacuate %{instance} \"%{name}\": %{details}") % {
-          :instance => ui_lookup(:table => 'vm_cloud'),
-          :name     => @record.name,
-          :details  => @record.is_available_now_error_message(:evacuate)}, :error)
-      end
-      params[:id] = @record.id.to_s # reset id in params for show
-      @record = nil
-      @sb[:action] = nil
-      replace_right_cell
-    end
-  end
-
   private
 
   def features
@@ -483,10 +80,5 @@ class VmCloudController < ApplicationController
 
   def skip_breadcrumb?
     breadcrumb_prohibited_for_action?
-  end
-
-  def get_error_message_from_fog(ex)
-    matched_message = ex.match(/message\\\": \\\"(.*)\\\", /)
-    matched_message ? matched_message[1] : ex
   end
 end
