+++ app/views/vm/show.html.haml
@@ -5,3 +5,13 @@
     = render :partial => "vm_common/right_size"
   - elsif @reconfigure
     = render :partial => "vm_common/reconfigure"
+  - elsif @resize
+    = render :partial => "vm_common/resize"
+  - elsif @live_migrate
+    = render :partial => "vm_common/live_migrate"
+  - elsif @evacuate
+    = render :partial => "vm_common/evacuate"
+  - elsif @attach
+    = render :partial => "vm_common/attach"
+  - elsif @detach
+    = render :partial => "vm_common/detach"
