+++ app/views/vm_common/_attach.html.haml
@@ -28,6 +28,25 @@
                   'paging_div_buttons_id'            => "angular_paging_div_buttons",
                   'paging_div_buttons_type'          => "Submit"}
 
+- unless @explorer
+  %table{:width => '100%'}
+    %tr
+      %td{:align => 'right'}
+        #buttons_on
+          = button_tag("Submit",
+                       :class        => "btn btn-primary",
+                       "ng-click"    => "submitClicked()",
+                       "ng-disabled" => "angularForm.$pristine || angularForm.$invalid",
+                       "ng-class"    => "{'btn-disabled': angularForm.$pristine || angularForm.$invalid}")
+          = button_tag("Reset",
+                       :class        => "btn btn-primary",
+                       "ng-click"    => "resetClicked()",
+                       "ng-disabled" => "angularForm.$pristine",
+                       "ng-class"    => "{'btn-disabled': angularForm.$pristine}")
+          = button_tag("Cancel",
+                        :class     => "btn btn-default",
+                        "ng-click" => "cancelClicked()")
+
 :javascript
-  ManageIQ.angular.app.value('vmCloudAttachFormId', '#{@vm.id}');
+  ManageIQ.angular.app.value('vmCloudAttachFormId', '#{@record.id}');
   miq_bootstrap('#form_div');
