+++ app/views/vm_common/_evacuate.html.haml
@@ -48,6 +48,21 @@
                   'paging_div_buttons_state_enabled' => true,
                   'paging_div_buttons_id'            => "angular_paging_div_buttons",
                   'paging_div_buttons_type'          => "Submit"}
+- unless @explorer
+  %table{:width => '100%'}
+    %tr
+      %td{:align => 'right'}
+        #buttons_on
+          = button_tag(t = _('Submit'),
+            :class   => "btn btn-primary",
+            :alt     => t,
+            :title   => t,
+            :onclick => "miqAjaxButton('#{url_for(:action => "evacuate_vm", :id => "#{@record.id}", :button => "submit")}');")
+          = button_tag(t = _('Cancel'),
+            :class   => "btn btn-default",
+            :alt     => t,
+            :title   => t,
+            :onclick => "miqAjaxButton('#{url_for(:action => "evacuate_vm", :id => "#{@record.id}", :button => "cancel")}');")
 
 :javascript
   ManageIQ.angular.app.value('vmCloudEvacuateFormId', '#{@record.id}');
