+++ app/views/vm_common/_resize.html.haml
@@ -25,3 +25,31 @@
       :javascript
         miqInitSelectPicker();
         miqSelectPickerEvent("flavor", "#{url}");
+- unless @explorer
+  %table{:width => "100%"}
+    %tr
+      %td{:align => 'right'}
+        #buttons_on{:style => "display:#{@changed ? "display" : "none"};"}
+          = button_tag(_("Submit"),
+                         :class   => 'btn btn-primary',
+                         :alt     => 'Submit',
+                         :title   => 'Submit',
+                         :onclick => "miqAjaxButton('#{url_for(:action => 'resize_vm',
+                                                               :id     => @edit[:vm_id],
+                                                               :button => "submit")}');")
+          = button_tag(_("Cancel"),
+                           :class   => 'btn btn-default',
+                           :alt     => t = _("Cancel Changes"),
+                           :title   => t,
+                           :onclick => "miqAjaxButton('#{url_for(:action => 'resize_vm',
+                                                                 :id     => @edit[:vm_id],
+                                                                 :button => "cancel")}');")
+        #buttons_off{:style => "display:#{@changed ? "none" : "display"};"}
+          = button_tag(_("Submit"), :class => "btn btn-primary btn-disabled")
+          = button_tag(_("Cancel"),
+                           :class   => 'btn btn-default',
+                           :alt     => t = _("Cancel Changes"),
+                           :title   => t,
+                           :onclick => "miqAjaxButton('#{url_for(:action => 'resize_vm',
+                                                                 :id     => @edit[:vm_id],
+                                                                 :button => "cancel")}');")
